+++ CHANGELOG.md
@@ -24,7 +24,7 @@ New features:
 * Added column 'Username' in 'Job Info' -> 'Hashes'.
 * Added column 'Username' in 'Hashes'.
 * Added dropdown menu to specify input format.
-* Added support for Telegram/Discord notifications.
+* Added support for E-mail/Telegram/Discord notifications.
 
 General improvements:
 * Improved/reimplemeted generator and scheduling logic. Better precision, more predictable workunit sizes. Default desired seconds per workunit changed to 600 seconds.